Complete single VM provisioning functions
