+++ central_service/ansible/roles/wait_for_ssh/tasks/main.yml
@@ -1,0 +1,7 @@
+---
+
+- name: Wait for port 22 to be ready
+  local_action: wait_for port=22 host="{{ inventory_hostname }}"  search_regex=OpenSSH
+
+- name: Wait for SSH SIGHUPs to stop
+  pause: seconds=15
